Extract Ckb.Address and Ckb.Types
