+++ src/Ckb.Address/Ckb.Address.csproj
@@ -7,8 +7,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Tippy.Util\Tippy.Util.csproj" />
     <ProjectReference Include="..\Ckb.Types\Ckb.Types.csproj" />
   </ItemGroup>
-
 </Project>
