+++ src/Ckb.Types/Ckb.Types.csproj
@@ -6,9 +6,4 @@
     <DebugType>embedded</DebugType>
   </PropertyGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\Tippy.Util\Tippy.Util.csproj" />
-    <ProjectReference Include="..\Ckb.Types\Ckb.Types.csproj" />
-  </ItemGroup>
-
 </Project>
